util: create data directories owner-only, from the chooser AND the core's first start
A new data directory inherited the umask default -- 0755 on macOS, whatever
%APPDATA% inherits on Windows. That is not an acceptable default for a directory
holding wallet.dat.

The hardening existed but sat where it could not run. It was in
ipc::Process::bind(), guarded on "did this call create the directory" -- and by
then GetDataDir() has already created it (ArgsManager::GetDataDirPath calls
fs::create_directories, and AppInit2 uses the datadir throughout). So the branch
never fired for the default address, and review finding 1.2 -- "the only remaining
barrier is an inherited datadir ACL that no code sets or checks" -- was not
actually closed for the directory itself.

util::CreateOwnerOnlyDirectory() creates and, only when it did the creating,
restricts: 0700 on POSIX, an owner+SYSTEM PROTECTED DACL on Windows, read back
and verified so a volume that stores no ACL is reported rather than trusted. It
is called from BOTH places a data directory can come into existence:

  * Intro::showIfNeeded (qt/intro.cpp) -- the monolithic first-run chooser, where
    the user picks the location;
  * ArgsManager::GetDataDirPath (util/system.cpp) -- everything else, and the one
    that matters for multiprocess, because the daemon starts first there (systemd
    unit, Windows scheduled task) and bitcoin.cpp:897 skips the chooser entirely
    under -multiprocess.

Covering only the chooser would have left the shipping multiprocess configuration
unprotected.

The Windows DACL machinery moves from ipc/process.cpp to util/ for this. It has
to: gridcoin_ipc is only built when ENABLE_MULTIPROCESS=ON, and the monolithic
build creates data directories too. ipc/process.cpp and ipc/handshake.cpp now use
the util implementation, so there is one copy of the security-relevant code
rather than two. gridcoin_util gains advapi32 on WIN32.

Create-only is preserved exactly: an existing directory is left as the operator
configured it, so a deliberately widened datadir is never silently re-tightened.
Failure to restrict a directory we just made is reported, not fatal -- it warns
in the core path (which runs before logging is up) and shows a dismissible
warning in the chooser, because the IPC socket and cookie carry their own
owner-only protection and fail closed independently.

Verified: a fresh default datadir and its network subdirectory are both created
0700. Circular-dependency lint still clean -- the new header depends only on
fs.h, and that ratchet is at zero.
